// portal/proto/isuxportal/resources/contestant_instance_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/resources/contestant_instance.proto (package isuxportal.proto.resources, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";
import type { Team } from "./team_pb";

/**
 * Describes the file isuxportal/resources/contestant_instance.proto.
 */
export declare const file_isuxportal_resources_contestant_instance: GenFile;

/**
 * @generated from message isuxportal.proto.resources.ContestantInstance
 */
export declare type ContestantInstance = Message<"isuxportal.proto.resources.ContestantInstance"> & {
  /**
   * @generated from field: int64 id = 7;
   */
  id: bigint;

  /**
   * @generated from field: string cloud_id = 1;
   */
  cloudId: string;

  /**
   * @generated from field: int64 team_id = 2;
   */
  teamId: bigint;

  /**
   * @generated from field: int64 number = 3;
   */
  number: bigint;

  /**
   * @generated from field: string public_ipv4_address = 4;
   */
  publicIpv4Address: string;

  /**
   * @generated from field: string private_ipv4_address = 5;
   */
  privateIpv4Address: string;

  /**
   * @generated from field: isuxportal.proto.resources.ContestantInstance.Status status = 6;
   */
  status: ContestantInstance_Status;

  /**
   * @generated from field: isuxportal.proto.resources.Team team = 16;
   */
  team?: Team;
};

/**
 * Describes the message isuxportal.proto.resources.ContestantInstance.
 * Use `create(ContestantInstanceSchema)` to create a new message.
 */
export declare const ContestantInstanceSchema: GenMessage<ContestantInstance>;

/**
 * @generated from enum isuxportal.proto.resources.ContestantInstance.Status
 */
export enum ContestantInstance_Status {
  /**
   * @generated from enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   * @generated from enum value: PENDING = 1;
   */
  PENDING = 1,

  /**
   * @generated from enum value: MODIFYING = 2;
   */
  MODIFYING = 2,

  /**
   * @generated from enum value: STOPPED = 3;
   */
  STOPPED = 3,

  /**
   * @generated from enum value: RUNNING = 4;
   */
  RUNNING = 4,

  /**
   * @generated from enum value: TERMINATED = 5;
   */
  TERMINATED = 5,
}

/**
 * Describes the enum isuxportal.proto.resources.ContestantInstance.Status.
 */
export declare const ContestantInstance_StatusSchema: GenEnum<ContestantInstance_Status>;